Manage my core bundles using bl
Buried in TextMate's source, there's [the source for a tool called `bl`](
https://github.com/textmate/textmate/blob/master/Applications/bl/src/bl.cc).
Talking to @sorbits, it seems he intended it mostly as a utility for use
when building TextMate, and so it isn't bundled in the app, and doesn't
really work for reading the state of the installed bundles. It still works
(and seems like it will continue to work) for installing, though. This
commit includes a pre-built copy of it to save me having to do that again.
Ordinarily, I wouldn't commit artefacts, but this seems like an
exception worth making.

This also installs a couple of bundles from GitHub (including my own
little one), and ignores the bundle directories more broadly.